Fix to allow import/use module from any folder
Load the shared library "ice_secp256k1.dll" or "ice_secp256k1.so" located in the same directory as the python module. By this fix the user can import the module from ANY directory using Python's CLI or IDE. Without this fix, the user can only import the module if the current working directory is the directory of this module.